LayoutTests/imported/w3c/web-platform-tests/reporting/same-origin-report-credentials.https.sub-expected.txt

CONSOLE MESSAGE: Refused to load https://localhost:9443/reporting/resources/fail.png because it does not appear in the img-src directive of the Content Security Policy.


PASS Reporting endpoints received credentials.
